Flask/login/flask-jwt-extended/requirements.txt
flask
bcrypt
pymongo
Flask-Bcrypt
python-dotenv
flask_sqlalchemy
Flask-JWT-Extended